backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-multipart==0.0.9
sqlmodel==0.0.22
pillow==10.4.0
opencv-python-headless==4.10.0.84
pytesseract==0.3.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psycopg[binary]==3.2.9
python-dotenv==1.0.1
httpx==0.27.0
prometheus-client==0.20.0
qrcode[pil]==7.4.2

